SSL自動更新情報反映

+++ app/Console/Commands/GetCmsSslInfo.php
@@ -19,7 +19,7 @@ class GetCmsSslInfo extends Command
      *
      * @var string
      */
-    protected $signature = 'app:get-cms-ssl-info';
+    protected $signature = 'app:get-cms-ssl-info {--test=false}';
 
     /**
      * The console command description.
@@ -33,7 +33,14 @@ class GetCmsSslInfo extends Command
      */
     public function handle()
     {
-        CmsSslInfo::truncate();
+        $test = $this->option('test');
+        if ($test) {
+            $this->info('test mode');
+        }
+
+        if (!$test) {
+            CmsSslInfo::truncate();
+        }
         $sslInfos = NotionReaderService::readCmsSslInfo();
         // NotionReaderService::updateCmsSslInfo($sslInfos);
 
@@ -41,7 +48,9 @@ class GetCmsSslInfo extends Command
             unset($sslInfo['page_id']);
             return $sslInfo;
         })->toArray();
-        CmsSslInfo::insert($sslInfos);
+        if (!$test) {
+            CmsSslInfo::insert($sslInfos);
+        }
 
         Log::info("CMSのSSL証明書の情報を取得しました。\n" . json_encode($sslInfos, JSON_PRETTY_PRINT));
 
@@ -49,25 +58,26 @@ class GetCmsSslInfo extends Command
         // SslInfo::truncate();
         // SslInfo::insert($sslInfos);
 
+        if (!$test) {
+            $today = Carbon::now();
+            $holidays = Yasumi::create('Japan', $today->format('Y'), 'ja_JP');
 
-        $today = Carbon::now();
-        $holidays = Yasumi::create('Japan', $today->format('Y'), 'ja_JP');
+            $isHoliday = $holidays->isHoliday($today);
+            $isWednesday = $today->dayOfWeek === Carbon::WEDNESDAY;
+            $shouldNotify = !$isHoliday && $isWednesday;
 
-        $isHoliday = $holidays->isHoliday($today);
-        $isWednesday = $today->dayOfWeek === Carbon::WEDNESDAY;
-        $shouldNotify = !$isHoliday && $isWednesday;
-
-        # send slack notification if --notify option is set
-        if ($shouldNotify) {
-            Log::info('Domain SSL 情報をSlackに通知します');
-            SlackService::sslInfo($sslInfos);
-        } else {
-            Log::info('Domain SSL 情報をSlackに通知しません', [
-                'today' => $today->format('Y-m-d'),
-                'isHoliday' => $isHoliday ? '○' : '×',
-                'isWednesday' => $isWednesday ? '○' : '×',
-                'shouldNotify' => $shouldNotify ? '○' : '×',
-            ]);
+            # send slack notification if --notify option is set
+            if ($shouldNotify) {
+                Log::info('Domain SSL 情報をSlackに通知します');
+                SlackService::sslInfo($sslInfos);
+            } else {
+                Log::info('Domain SSL 情報をSlackに通知しません', [
+                    'today' => $today->format('Y-m-d'),
+                    'isHoliday' => $isHoliday ? '○' : '×',
+                    'isWednesday' => $isWednesday ? '○' : '×',
+                    'shouldNotify' => $shouldNotify ? '○' : '×',
+                ]);
+            }
         }
     }
 }
